requirements.txt
absl-py==2.2.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
Bottleneck==1.4.2
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dacite==1.9.2
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.1
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.15.4
flatbuffers==25.2.10
fonttools==4.55.5
fqdn==1.5.1
gast==0.6.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
googletrans==4.0.0rc1
grpcio==1.71.0
h11==0.9.0
h2==3.2.0
h5py==3.13.0
hpack==3.0.0
hstspreload==2025.1.1
htmlmin==0.1.12
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ImageHash==4.3.1
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
lxml==5.3.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
mistune==3.1.0
ml_dtypes==0.5.1
mlxtend==0.23.4
multimethod==1.4
namex==0.0.8
narwhals==1.30.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.1.3
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandas-profiling==3.0.0
pandas-stubs==2.2.3.241126
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
phik==0.12.4
pillow==11.1.0
platformdirs==4.2.2
playsound==1.2.2
plotly==6.0.0
polars==1.23.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
protobuf==5.29.4
psutil==6.1.0
pure_eval==0.2.3
pyarrow==19.0.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.25.5
pyparsing==3.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-json-logger==3.2.1
pytz==2024.2
PyWavelets==1.8.0
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==72.1.0
six==1.16.0
skimpy==0.0.18
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.2
sqlparse==0.5.1
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.44.1
tangled-up-in-unicode==0.1.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.0.1
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.2
types-python-dateutil==2.9.0.20241206
types-pytz==2025.1.0.20250204
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.3.0
virtualenv==20.26.3
visions==0.7.1
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
word2number==1.1
wordcloud==1.9.4
wrapt==1.17.2
ydata-profiling==4.12.2
